% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psuedocohort.R
\name{pseudocohort}
\alias{pseudocohort}
\title{Generate Psuedo Longitudinal Cohort}
\usage{
pseudocohort(longmin, longmax, latmin, latmax, numberinsample)
}
\arguments{
\item{longmin}{minimum longitude of bounding box}

\item{longmax}{maximum longitude of bounding box}

\item{latmin}{minimum latitude of bounding box}

\item{latmax}{maximum longitude of bounding box}

\item{numberofaddresses}{How many addresses to try and sample}
}
\value{
A dataframe of addresses in bounding box
}
\description{
This function generates random Open Street Map addresses for a given bounding box
}
