% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generatepseudocohort.R
\name{psuedocohort}
\alias{psuedocohort}
\title{Generate longitudinal pseudo cohort}
\usage{
psuedocohort(birthyear, longmin, longmax, latmin, latmax)
}
\arguments{
\item{birthyear}{year of birth for cohort}

\item{longmax}{maximum longitude of bounding box}

\item{latmin}{minimum latitude of bounding box}

\item{latmax}{maximum longitude of bounding box}
}
\value{
A cohort of participants with longitudinal address history
}
\description{
This function generates a longitudinal cohort of participants
with random Open Street Map addresses movement
}
